my_django_app/requirements.txt
asgiref==3.3.4
dj-database-url==0.5.0
Django==3.2
django-ninja==0.12.3
psycopg2==2.8.6
psycopg2-binary==2.8.6
pybase64==1.1.4
pydantic==1.8.1
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
unicorn==1.0.2
whitenoise==5.2.0
